+++ src/main/java/com/example/capstone_be/dto/tour/TourViewDto.java
@@ -29,9 +29,9 @@ public class TourViewDto implements Serializable {
 
     private String working;
 
-    private Double latitude;
+    private Float latitude;
 
-    private Double longitude;
+    private Float longitude;
 
     private String destination;
